requirements.txt
beautifulsoup4
chromadb
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
langgraph
lxml
python-dotenv
requests
sentence-transformers
streamlit
tiktoken
